+++ moysklad/demand.go
@@ -30,7 +30,7 @@ type Demand struct {
 	OrganizationAccount     *AgentAccount              `json:"organizationAccount,omitempty"`
 	Overhead                *Overhead                  `json:"overhead,omitempty"`
 	Owner                   *Employee                  `json:"owner,omitempty"`
-	PayedSum                *Decimal                   `json:"payedSum,omitempty"`
+	PayedSum                *float64                   `json:"payedSum,omitempty"`
 	Positions               *Positions[DemandPosition] `json:"positions,omitempty"`
 	Printed                 *bool                      `json:"printed,omitempty"`
 	Project                 *Project                   `json:"project,omitempty"`
@@ -42,12 +42,12 @@ type Demand struct {
 	ShipmentAddressFull     *Address                   `json:"shipmentAddressFull,omitempty"`
 	State                   *State                     `json:"state,omitempty"`
 	Store                   *Store                     `json:"store,omitempty"`
-	Sum                     *Decimal                   `json:"sum,omitempty"`
+	Sum                     *float64                   `json:"sum,omitempty"`
 	SyncID                  *uuid.UUID                 `json:"syncId,omitempty"`
 	Updated                 *Timestamp                 `json:"updated,omitempty"`
 	VatEnabled              *bool                      `json:"vatEnabled,omitempty"`
 	VatIncluded             *bool                      `json:"vatIncluded,omitempty"`
-	VatSum                  *Decimal                   `json:"vatSum,omitempty"`
+	VatSum                  *float64                   `json:"vatSum,omitempty"`
 	CustomerOrder           *CustomerOrder             `json:"customerOrder,omitempty"`
 	FactureOut              *FactureOut                `json:"factureOut,omitempty"`
 	Returns                 *SalesReturns              `json:"returns,omitempty"`
@@ -88,16 +88,16 @@ type DemandPosition struct {
 	AccountID         *uuid.UUID          `json:"accountId,omitempty"`          // ID учетной записи
 	Assortment        *AssortmentPosition `json:"assortment,omitempty"`         // Метаданные товара/услуги/серии/модификации/комплекта, которую представляет собой позиция
 	Cost              *int                `json:"cost,omitempty"`               // Себестоимость (только для услуг)
-	Discount          *Decimal            `json:"discount,omitempty"`           // Процент скидки или наценки. Наценка указывается отрицательным числом, т.е. -10 создаст наценку в 10%
+	Discount          *float64            `json:"discount,omitempty"`           // Процент скидки или наценки. Наценка указывается отрицательным числом, т.е. -10 создаст наценку в 10%
 	ID                *uuid.UUID          `json:"id,omitempty"`                 // ID сущности
 	Pack              *Pack               `json:"pack,omitempty"`               // Упаковка Товара
-	Price             *Decimal            `json:"price,omitempty"`              // Цена товара/услуги в копейках
+	Price             *float64            `json:"price,omitempty"`              // Цена товара/услуги в копейках
 	Quantity          *float64            `json:"quantity,omitempty"`           // Количество товаров/услуг данного вида в позиции. Если позиция - товар, у которого включен учет по серийным номерам, то значение в этом поле всегда будет равно количеству серийных номеров для данной позиции в документе.
 	Slot              *Slot               `json:"slot,omitempty"`               // Ячейка на складе
 	Things            *Things             `json:"things,omitempty"`             // Серийные номера. Значение данного атрибута игнорируется, если товар позиции не находится на серийном учете. В ином случае количество товаров в позиции будет равно количеству серийных номеров, переданных в значении атрибута
 	TrackingCodes     *TrackingCodes      `json:"trackingCodes,omitempty"`      // Коды маркировки товаров и транспортных упаковок
 	TrackingCodes1162 *TrackingCodes      `json:"trackingCodes_1162,omitempty"` // Коды маркировки товаров в формате тега 1162
-	Overhead          *Decimal            `json:"overhead,omitempty"`           // Накладные расходы. Если Позиции Отгрузки не заданы, то накладные расходы нельзя задать
+	Overhead          *float64            `json:"overhead,omitempty"`           // Накладные расходы. Если Позиции Отгрузки не заданы, то накладные расходы нельзя задать
 	Vat               *int                `json:"vat,omitempty"`                // НДС, которым облагается текущая позиция
 	VatEnabled        *bool               `json:"vatEnabled,omitempty"`         // Включен ли НДС для позиции. С помощью этого флага для позиции можно выставлять НДС = 0 или НДС = "без НДС". (vat = 0, vatEnabled = false) -> vat = "без НДС", (vat = 0, vatEnabled = true) -> vat = 0%.
 	Stock             *Stock              `json:"stock,omitempty"`              // Остатки и себестоимость `?fields=stock&expand=positions`
